platform-enterprise-python/libs/platform-observability/pyproject.toml
[project]
name = "platform-observability"
version = "1.0.0"
description = "Platform observability - logging, metrics, tracing"
requires-python = ">=3.12"
dependencies = [
    "structlog>=24.4.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
    "opentelemetry-instrumentation-redis>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "prometheus-client>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/platform_observability"]
